+++ src/com/library/core/Book.kt
@@ -1,0 +1,7 @@
+package com.library.core
+
+data class Book(
+    val id: String,
+    val title: String,
+    val year: String,
+)
