+++ src/com/library/db/entities/BookEntity.kt
@@ -1,0 +1,9 @@
+package com.library.db.entities
+
+import java.io.Serializable
+
+data class BookEntity(
+    val id: String,
+    val title: String,
+    val year: String,
+) : Serializable
