+++ src/com/library/db/entities/ClientEntity.kt
@@ -1,0 +1,9 @@
+package com.library.db.entities
+
+import java.io.Serializable
+
+data class ClientEntity(
+    val id: String,
+    val name: String,
+    val surname: String,
+) : Serializable
