+++ src/com/library/db/entities/OwnershipEntity.kt
@@ -1,0 +1,8 @@
+package com.library.db.entities
+
+import java.io.Serializable
+
+data class OwnershipEntity(
+    val client: ClientEntity,
+    val booksInUse: MutableSet<BookEntity>,
+) : Serializable
